+++ mdspan/include/nabla/elementwise_expr.hpp
@@ -1,7 +1,7 @@
 #include <functional>
 #include <utility>
 #include <tuple>
-#include "nabla/traits.hpp"
+#include "nabla/concepts.hpp"
 
 // TODO: enforce invariants e.g. rank, dimensions, fp type
 
@@ -11,19 +11,8 @@
 
 namespace nabla {
 
-template <typename T>
-    requires IsExprCompatible<T>
-class ExprStorage {
-    T _expr;
-    public:
-        ExprStorage(const T& expr) : _expr(expr) {}
-        ExprStorage(T&& expr) : _expr(std::move(expr)) {}
-        T::index_type size() const { return _expr.size(); }
-        const T& get() const { return _expr; }
-        T& get() { return _expr; }
-};
-
 template <typename Op, typename... Inputs>
+    requires (IsElementwiseExprCompatible<Inputs> && ...)
 class ExprElementWiseOp;
 
 // base case: collect_leaf_ptrs for a leaf nodes
@@ -33,27 +22,19 @@ auto collect_leaf_ptrs(T& leaf) {
     return std::tuple<T*>{&leaf};
 }
 
-// returns inputs of an expression in left-to-right depth-first order
-//template <typename Op, typename... Inputs>
-//auto collect_leaf_ptrs(ExprElementWiseOp<Op, Inputs...>& expr) {
-//    return std::tuple_cat(collect_leaf_ptrs(std::get<ExprStorage<Inputs>>(expr._inputs).get())...);
-//}
-
-template <typename Op, typename... Inputs, std::size_t... Is>
-auto collect_leaf_ptrs_impl(ExprElementWiseOp<Op, Inputs...>& expr, std::index_sequence<Is...>) {
-    return std::tuple_cat(collect_leaf_ptrs(std::get<Is>(expr._inputs).get())...);
-}
-
 template <typename Op, typename... Inputs>
 auto collect_leaf_ptrs(ExprElementWiseOp<Op, Inputs...>& expr) {
-    return collect_leaf_ptrs_impl(expr, std::index_sequence_for<Inputs...>{});
+    return [&]<std::size_t... Is>(std::index_sequence<Is...>) {
+        return std::tuple_cat(collect_leaf_ptrs(std::get<Is>(expr._inputs))...);
+    }(std::index_sequence_for<Inputs...>{});
 }
 
 // N-ary expression template
 template <typename Op, typename... Inputs>
+    requires (IsElementwiseExprCompatible<Inputs> && ...)
 class ExprElementWiseOp : public ElementwiseExprTag {
     Op _op;
-    std::tuple<ExprStorage<Inputs>...> _inputs;
+    std::tuple<Inputs...> _inputs;
     using input1_t = std::tuple_element_t<0, std::tuple<Inputs...>>;
 
     public:
@@ -62,7 +43,7 @@ class ExprElementWiseOp : public ElementwiseExprTag {
         // Member types
         //
         using op_type = Op;
-        using inputs_type = std::tuple<ExprStorage<Inputs>...>;
+        using inputs_type = std::tuple<Inputs...>;
         using extents_type = typename input1_t::extents_type;
         using index_type = typename input1_t::index_type;
         using coord_type = typename input1_t::coord_type;
@@ -78,9 +59,9 @@ class ExprElementWiseOp : public ElementwiseExprTag {
         //
         static constexpr input1_t::rank_type rank() noexcept { return input1_t::rank(); }
 
-        constexpr index_type extent(rank_type r) const noexcept { return std::get<0>(_inputs).get().extent(r); }
-        constexpr extents_type extents() const noexcept { return std::get<0>(_inputs).get().extents(); }
-        constexpr index_type size() const noexcept { return std::get<0>(_inputs).get().size(); }
+        constexpr index_type extent(rank_type r) const noexcept { return std::get<0>(_inputs).extent(r); }
+        constexpr extents_type extents() const noexcept { return std::get<0>(_inputs).extents(); }
+        constexpr index_type size() const noexcept { return std::get<0>(_inputs).size(); }
 
         //
         // Constructors
@@ -95,7 +76,7 @@ class ExprElementWiseOp : public ElementwiseExprTag {
         auto operator[](size_t i) const {
             return std::apply(
                 [&](auto const&... ins) {
-                    return _op(ins.get()[i]...);
+                    return _op(ins[i]...);
                 }, _inputs);
         }
 
@@ -104,12 +85,12 @@ class ExprElementWiseOp : public ElementwiseExprTag {
         auto operator()(Args... args) const {
             return std::apply(
                 [&](auto const&... ins) {
-                    return _op(ins.get()(args...)...);
+                    return _op(ins(args...)...);
                 }, _inputs);
         }
 
-        template <typename Op_, typename... Inputs_, std::size_t... Is>
-        friend auto collect_leaf_ptrs_impl(ExprElementWiseOp<Op_, Inputs_...>& expr, std::index_sequence<Is...>);
+        template <typename Op_, typename... Inputs_>
+        friend auto collect_leaf_ptrs(ExprElementWiseOp<Op_, Inputs_...>& expr);
 
         auto inputs() {
             return collect_leaf_ptrs(*this);
@@ -119,9 +100,9 @@ class ExprElementWiseOp : public ElementwiseExprTag {
 //helper to decay input types (e.g. const T& -> T)
 template <typename Op, typename... Inputs>
 auto make_expr_element_wise_op(Op&& op, Inputs&&... inputs) {
-    return ExprElementWiseOp< std::decay_t<Op>, std::decay_t<Inputs>... >(
+    return ExprElementWiseOp< std::decay_t<Op>, std::decay_t<Inputs>... >{
         std::forward<Op>(op), std::forward<Inputs>(inputs)...
-    );
+    };
 }
 
 // Operator overloads: Arithmetic operations (+ - * / % -)
