--- mdspan/include/nabla/traits.hpp
@@ -1,2 +1,0 @@
-#include "nabla/traits/tensor.hpp"
-#include "nabla/traits/expr.hpp"
